% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcorr.R
\name{distcorr}
\alias{distcorr}
\title{Distance Correlation}
\usage{
distcorr(x, y)
}
\arguments{
\item{x, y}{numeric vectors of data values}
}
\description{
Provides pairwise correlation via distance covariance normalized by distance standard deviation. Allows for non-linear dependencies.
}
\examples{
# Sample data
a <- rnorm(25, 80, 35)
b <- rnorm(25, 100, 50)

mlf::distcorr(a, b)
}
\references{
Székely GJ, Rizzo ML, Bakirov NK. Measuring and testing dependence by correlation of distances. Ann Stat. 2007. 35(6):2769-2794.
}
