% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(x, bins)
}
\arguments{
\item{x}{numeric or discrete data vector}

\item{bins}{specify number of bins if numeric or integer data class.}
}
\description{
Estimates uncertainty in univariate probability distribution.
}
\examples{
# Sample numeric vector
a <- rnorm(25, 80, 35)
mlf::entropy(a, bins = 2)

# Sample discrete vector
b <- as.factor(c(1,1,1,2))
mlf::entropy(b)
}
