% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bias.R
\name{get_bias}
\alias{get_bias}
\title{Bias}
\usage{
get_bias(truth, estimate)
}
\arguments{
\item{truth}{test data vector or baseline accuracy to test against.}

\item{estimate}{predicted vector}
}
\description{
Estimates squared bias by decomposing model prediction error.
}
\examples{
# Sample data
test <- rnorm(25, 80, 35)
predicted <- rnorm(25, 80, 50)

mlf::get_bias(test, predicted)
}
