% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mse.R
\name{get_mse}
\alias{get_mse}
\title{Mean Squared Error}
\usage{
get_mse(truth, estimate)
}
\arguments{
\item{truth}{test data vector or baseline accuracy to test against.}

\item{estimate}{predicted vector}
}
\description{
Estimates mean squared error from model predictions.
}
\examples{
# Sample data
test <- rnorm(25, 80, 35)
predicted <- rnorm(25, 80, 50)

mlf::get_mse(test, predicted)
}
