% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var.R
\name{get_var}
\alias{get_var}
\title{Variance}
\usage{
get_var(estimate)
}
\arguments{
\item{estimate}{predicted vector}
}
\description{
Estimates squared variance by decomposing model prediction error.
}
\examples{
# Sample data
test <- rnorm(25, 80, 35)
predicted <- rnorm(25, 80, 50)

mlf::get_var(predicted)
}
