% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic.R
\name{mic}
\alias{mic}
\title{Maximal Information Criterion}
\usage{
mic(x, y)
}
\arguments{
\item{x, y}{numeric or discrete data vectors}
}
\description{
Information-theoretic approach for detecting non-linear pairwise dependencies. Employs heuristic discretization to achieve highest normalized mutual information.
}
\examples{
# Sample data
a <- rnorm(25, 80, 35)
b <- rnorm(25, 100, 50)

mlf::mic(a, b)
}
\references{
Reshef DN, Reshef YA, Finucane HK, Grossman SR, McVean G, Turnbaugh PJ, Lander ES, Mitzenmacher M, Sabeti PC. Detecting novel associations in large data sets. Science. 2011. 334(6062):1518-1524.
}
